refactor: migrate chat link functionality to chat hash
